add context#StatusCodeNotSuccessful for customize even the most customized clients that are not compatible with the standards and fix the `SPA` if static file serve handlers are passed as its `AssetHandler` as reported at the chat.iris-go.com
